Expand the project description. Leave mention of forking to Credits section.
The text here is actually the description that I use in the RPM package.
It's maybe a bit wordy compared to the understated description we had
before, but I thought it might be helpful for browsing users who don't
already know what the Legacy Tray even is.

Also, I figure that, since the Credits section at the end thanks the
author of the original TopIcons extension, mentioning the fork at the
beginning too is redundant.